# sql: don't modify select expressions in place