We currently resolve qnames in expressions, modifying them in-place. This leads
to some hacky code to correctly handle auto-transaction retries, and this scheme
won't work at all when we will split filtering expressions.

We change this by making a deep-copy of the expression before qname resolution.
If needed, in the future we could improve the visitor model to allow us to make
copies only of the nodes that change.

Changing prepared update code to look at the selectNode render target
expressions instead of relying on the in-place modification (it's actually
easier this way). @mjibson

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/4418)
<!-- Reviewable:end -->