@RaduBerinde 2016-02-16
An alternative to deep copying would be to change the way we use qvalues a bit: instead of replacing QualifiedName nodes in the tree, we could add a `resolved Expr`field in QualifiedName. This field would be set during qname resolution (and it wouldn't matter if it was already set before). Other than these fields, the tree wouldn't have to change for qvalue resolving. Let me know what you think of this idea and if you think it would make things cleaner in general.

---







*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAg-MsIfa18aX2oYSKL:1781144808)*
<!-- Sent from Reviewable.io -->