@mjibson 2016-02-17
I don't understand why we can ignore the expression type (i.e., ignore if it's a tuple or not). I also don't understand why we only need to look at `tableDesc.allColumnsSelector()` (`targets` when the length is taken) instead of `n.Exprs`. Could you explain that?

---

Reviewed 1 of 3 files at r1.
Review status: 1 of 3 files reviewed at latest revision, all discussions resolved.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAgznseBA_wWjgho6WB:-1369194772)*
<!-- Sent from Reviewable.io -->