@petermattis 2016-02-17
I agree that the current setup where we mutate the expressions is problematic and error prone. On the other hand, making a full deep-copy of the expression seems excessive. I like the suggestion to make a copy of only the nodes that change. I'm a little worried that enhancing `QualifiedName` to contain a `resolved Expr` isn't a full solution as a schema change could change name resolution.