@tamird 2016-02-17
Would be good to see before/after benchmark numbers.

---

Reviewed 3 of 3 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[sql/parser/deep_copy.go, line 150 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-KAg-phfRfdE4pRgn3vt:-KAg-phfRfdE4pRgn3vu:-426685668) ([raw file](https://github.com/cockroachdb/cockroach/blob/bb6efa1c2a5ae059613cdbe4a121c579b78d1451/sql/parser/deep_copy.go#L150)):*
Can you add a test that hits this in deep_copy_test?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAhKbr_xSxAaPyIgomm:72632063)*
<!-- Sent from Reviewable.io -->