@RaduBerinde 2016-02-17
@mjibson we are looking at all targets *except* `allColumnsSelector` - these are the  targets created from `n.Expr` in the loop above which has the exact same logic as the loop we are replacing (tuple vs dtuple vs other). Instead of going through the same loop logic again, we just look at the targets that have been created in the previous loop. I'll add a comment.

@petermattis can you expand on the problem around schema change? The `resolved Expr` inside a `QualifiedName` would be set to a `*qvalue`, and we would never reuse a previously computed value - any time we `resolveQValues` we ignore any `resolved Expr` already there. We can also have a visitor that "cleans up" (sets to nil) all `resolved` values to make sure they don't inadvertently show up.

---







*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAjUNgdlX_Nddbn4C8X:770818428)*
<!-- Sent from Reviewable.io -->