@petermattis 2016-02-17
Ok, if you're never using the `resolved Expr` between executions, that seems safe. 

---

Review status: all files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAjrpChy-P1ZZDgrvQL:1529642437)*
<!-- Sent from Reviewable.io -->