@andreimatei 2016-02-17
I'm not sure I understand why the prepare code needs to change?

---

Review status: all files reviewed at latest revision, 5 unresolved discussions.

---

*[sql/select_qvalue.go, line 157 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-KAkD-im_dknebBfBhdG:-KAkD-im_dknebBfBhdH:1846998525) ([raw file](https://github.com/cockroachdb/cockroach/blob/bb6efa1c2a5ae059613cdbe4a121c579b78d1451/sql/select_qvalue.go#L157)):*
i'm glad we're getting rid of this code that I wasn't even convinced we needed

---

*[sql/select_qvalue.go, line 160 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-KAkCh0qmVOT8-7w8Z5C:-KAkCh0qmVOT8-7w8Z5D:-231215121) ([raw file](https://github.com/cockroachdb/cockroach/blob/bb6efa1c2a5ae059613cdbe4a121c579b78d1451/sql/select_qvalue.go#L160)):*
nit: move this comment as the panic message

---

*[sql/select_qvalue.go, line 226 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-KAk2yfJBVkZx6y4pCkc:-KAk2yfJBVkZx6y4pCkd:1906078488) ([raw file](https://github.com/cockroachdb/cockroach/blob/bb6efa1c2a5ae059613cdbe4a121c579b78d1451/sql/select_qvalue.go#L226)):*
nit: retrying transactions in general, not just "auto"

---

*[sql/update.go, line 96 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-KAk78uBbFmM3GpwoAo4:-KAk78uCQGeNJg39iyrj:-584365348) ([raw file](https://github.com/cockroachdb/cockroach/blob/bb6efa1c2a5ae059613cdbe4a121c579b78d1451/sql/update.go#L96)):*
selecting all the columns here is pessimistic, right? We only need to select the cols referenced in update expressions?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAkDMnSEauFaq9UGhAA:1265179101)*
<!-- Sent from Reviewable.io -->