@RaduBerinde 2016-02-17
^The prepare code is relying on the input expressions being changed in-place. If you read it carefully what it does is it sets up the Select after which we can "magically" call `TypeCheck` on the same expressions..

I will give the `resolved` inside `QualifiedName` idea a try and see how it looks.

---







*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAkJznDZW3zmkutT8gH:1716374066)*
<!-- Sent from Reviewable.io -->