@RaduBerinde 2016-02-17
I added a benchmark with a `SELECT` with a bunch of expressions (against a small table) and I don't see a regression.

I will still try the alternative approach I mentioned above to see if it makes things cleaner.

```
name                 old time/op  new time/op  delta
Select1_Cockroach-4   104µs ±10%   106µs ±34%   ~     (p=0.353 n=10+10)
Select2_Cockroach-4  1.21ms ± 8%  1.19ms ± 9%   ~       (p=0.370 n=8+9)
```

---

Review status: 0 of 4 files reviewed at latest revision, 5 unresolved discussions.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAkXShI1ZNqOuHW7h1n:217423741)*
<!-- Sent from Reviewable.io -->