@RaduBerinde 2016-02-17
Review status: 0 of 4 files reviewed at latest revision, 5 unresolved discussions.

---

*[sql/update.go, line 96 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-KAk78uBbFmM3GpwoAo4:-KAkY8qYyDekG6iBTIii:1818820584) ([raw file](https://github.com/cockroachdb/cockroach/blob/bb6efa1c2a5ae059613cdbe4a121c579b78d1451/sql/update.go#L96)):*
I don't know this code well. At a first glance we might also need some columns to encode keys.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418)*
<!-- Sent from Reviewable.io -->