@RaduBerinde 2016-02-17
By the way, I think we have a related problem with subqueries. The subquery visitor will replace subquery parts of expressions with DTuples or whatever. Upon retry of the transaction, the expression has been changed and we will not process the subqueries again. If the tables used by the subquery changed in the meantime, the results might be inconsistent.. Is this right or am I missing something?

---

Review status: 0 of 4 files reviewed at latest revision, 5 unresolved discussions.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4418#-:-KAkiHEoGqFVu9AL9mt0:-1142897861)*
<!-- Sent from Reviewable.io -->